Refactor AssetsRepository into focused services
Peel single-responsibility concerns out of AssetsRepository (#470),
mirroring the iOS service split. The repository stays the entry point
and delegates:

- AssetsAvailabilityService — buy/sell/swap availability
- CurrencyRatesService — currency rate conversion and lookup
- AssetsSearchService — asset and swap search
- RecentAssetsService — recent assets
- TransactionPostProcessingService — post-transaction balance refresh
  and stake/NFT resync, now started from App instead of the repo init

Behavior is unchanged; existing public methods keep delegating.
