+++ android/app/src/main/kotlin/com/gemwallet/android/App.kt
@@ -12,6 +12,7 @@ import coil3.memory.MemoryCache
 import coil3.svg.SvgDecoder
 import com.gemwallet.android.application.perpetual.coordinators.SyncPerpetualPositions
 import com.gemwallet.android.application.transactions.coordinators.GetTransactions
+import com.gemwallet.android.data.repositories.assets.TransactionPostProcessingService
 import com.gemwallet.android.data.repositories.stream.StreamObserverService
 import dagger.hilt.android.HiltAndroidApp
 import javax.inject.Inject
@@ -25,6 +26,8 @@ class App : Application(), SingletonImageLoader.Factory, Application.ActivityLif
     lateinit var syncPerpetualPositions: SyncPerpetualPositions
     @Inject
     lateinit var getTransactions: GetTransactions
+    @Inject
+    lateinit var transactionPostProcessingService: TransactionPostProcessingService
 
     override fun onCreate() {
         super.onCreate()
